BackEnd/requirements.txt
aiosmtpd==1.4.6
amqp==5.3.1
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
atpublic==5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
binaryornot==0.4.4
build==1.2.2.post1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.2
daphne==4.1.2
Django==5.2
django-axes==7.0.2
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-crum==0.7.9
django-filter==25.1
django-redis==5.4.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
googlemaps==4.10.0
h11==0.16.0
hiredis==3.1.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pip-check==3.1
pip-tools==7.4.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-gettext==5.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.6
selenium==4.31.0
serpapi==0.1.5
service-identity==24.2.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stripe==12.0.1
tenacity==9.1.2
terminaltables==3.1.10
text-unidecode==1.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
wsproto==1.2.0
zope.interface==7.2
